igscraper/requirements.txt
fastapi
uvicorn
psycopg2-binary
alembic
sqlalchemy
pydantic
ensta
python-decouple
httpx
